requirements.txt
transformers
datasets
peft
trl
pyyaml
torch
wandb
tensorboardX
pytest